Revert "fix: validate configs and secrets against the schema before setting (#1858)"
This reverts commit f4cb64adaae99fd2ab1739a0d84c2393f794ce9e.

# Conflicts:
#	backend/controller/admin/testdata/go/ftl-project-dr.toml
