+++ backend/controller/admin/admin_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/alecthomas/types/optional"
 
 	ftlv1 "github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1"
-	"github.com/TBD54566975/ftl/backend/schema"
 	cf "github.com/TBD54566975/ftl/common/configuration"
 	"github.com/TBD54566975/ftl/internal/log"
 )
@@ -32,7 +31,7 @@ func TestAdminService(t *testing.T) {
 			cf.InlineProvider[cf.Secrets]{},
 		})
 	assert.NoError(t, err)
-	admin := NewAdminService(cm, sm, &diskSchemaRetriever{})
+	admin := NewAdminService(cm, sm)
 	assert.NotZero(t, admin)
 
 	expectedEnvarValue, err := json.MarshalIndent(map[string]string{"bar": "barfoo"}, "", "  ")
@@ -140,131 +139,4 @@ func testAdminSecrets(
 		assert.NoError(t, err)
 		assert.Equal(t, entry.Value, string(resp.Msg.Value))
 	}
-}
-
-var testSchema = schema.MustValidate(&schema.Schema{
-	Modules: []*schema.Module{
-		{
-			Name:     "batmobile",
-			Comments: []string{"A batmobile comment"},
-			Decls: []schema.Decl{
-				&schema.Secret{
-					Comments: []string{"top secret"},
-					Name:     "owner",
-					Type:     &schema.String{},
-				},
-				&schema.Secret{
-					Comments: []string{"ultra secret"},
-					Name:     "horsepower",
-					Type:     &schema.Int{},
-				},
-				&schema.Config{
-					Comments: []string{"car color"},
-					Name:     "color",
-					Type:     &schema.Ref{Module: "batmobile", Name: "Color"},
-				},
-				&schema.Config{
-					Comments: []string{"car capacity"},
-					Name:     "capacity",
-					Type:     &schema.Ref{Module: "batmobile", Name: "Capacity"},
-				},
-				&schema.Enum{
-					Comments: []string{"Car colors"},
-					Name:     "Color",
-					Type:     &schema.String{},
-					Variants: []*schema.EnumVariant{
-						{Name: "Black", Value: &schema.StringValue{Value: "Black"}},
-						{Name: "Blue", Value: &schema.StringValue{Value: "Blue"}},
-						{Name: "Green", Value: &schema.StringValue{Value: "Green"}},
-					},
-				},
-				&schema.Enum{
-					Comments: []string{"Car capacities"},
-					Name:     "Capacity",
-					Type:     &schema.Int{},
-					Variants: []*schema.EnumVariant{
-						{Name: "One", Value: &schema.IntValue{Value: int(1)}},
-						{Name: "Two", Value: &schema.IntValue{Value: int(2)}},
-						{Name: "Four", Value: &schema.IntValue{Value: int(4)}},
-					},
-				},
-			},
-		},
-	},
-})
-
-type mockSchemaRetriever struct {
-}
-
-func (d *mockSchemaRetriever) GetActiveSchema(ctx context.Context) (*schema.Schema, error) {
-	return testSchema, nil
-}
-
-func TestAdminValidation(t *testing.T) {
-	config := tempConfigPath(t, "testdata/ftl-project.toml", "admin")
-	ctx := log.ContextWithNewDefaultLogger(context.Background())
-
-	cm, err := cf.NewConfigurationManager(ctx, cf.ProjectConfigResolver[cf.Configuration]{Config: config})
-	assert.NoError(t, err)
-
-	sm, err := cf.New(ctx,
-		cf.ProjectConfigResolver[cf.Secrets]{Config: config},
-		[]cf.Provider[cf.Secrets]{
-			cf.EnvarProvider[cf.Secrets]{},
-			cf.InlineProvider[cf.Secrets]{},
-		})
-	assert.NoError(t, err)
-	admin := NewAdminService(cm, sm, &mockSchemaRetriever{})
-	assert.NotZero(t, admin)
-
-	testSetConfig(t, ctx, admin, "batmobile", "color", "Black", "")
-	testSetConfig(t, ctx, admin, "batmobile", "color", "Red", "JSON validation failed: Red is not a valid variant of enum batmobile.Color")
-	testSetConfig(t, ctx, admin, "batmobile", "capacity", 2, "")
-	testSetConfig(t, ctx, admin, "batmobile", "capacity", 3, "JSON validation failed: %!s(float64=3) is not a valid variant of enum batmobile.Capacity")
-
-	testSetSecret(t, ctx, admin, "batmobile", "owner", "Bruce Wayne", "")
-	testSetSecret(t, ctx, admin, "batmobile", "owner", 99, "JSON validation failed: owner has wrong type, expected String found float64")
-	testSetSecret(t, ctx, admin, "batmobile", "horsepower", 1000, "")
-	testSetSecret(t, ctx, admin, "batmobile", "horsepower", "thousand", "JSON validation failed: horsepower has wrong type, expected Int found string")
-
-	testSetConfig(t, ctx, admin, "", "city", "Gotham", "")
-	testSetSecret(t, ctx, admin, "", "universe", "DC", "")
-}
-
-// nolint
-func testSetConfig(t testing.TB, ctx context.Context, admin *AdminService, module string, name string, jsonVal any, expectedError string) {
-	t.Helper()
-	buffer, err := json.Marshal(jsonVal)
-	assert.NoError(t, err)
-
-	configRef := &ftlv1.ConfigRef{Name: name}
-	if module != "" {
-		configRef.Module = &module
-	}
-
-	_, err = admin.ConfigSet(ctx, connect.NewRequest(&ftlv1.SetConfigRequest{
-		Provider: ftlv1.ConfigProvider_CONFIG_INLINE.Enum(),
-		Ref:      configRef,
-		Value:    buffer,
-	}))
-	assert.EqualError(t, err, expectedError)
-}
-
-// nolint
-func testSetSecret(t testing.TB, ctx context.Context, admin *AdminService, module string, name string, jsonVal any, expectedError string) {
-	t.Helper()
-	buffer, err := json.Marshal(jsonVal)
-	assert.NoError(t, err)
-
-	configRef := &ftlv1.ConfigRef{Name: name}
-	if module != "" {
-		configRef.Module = &module
-	}
-
-	_, err = admin.SecretSet(ctx, connect.NewRequest(&ftlv1.SetSecretRequest{
-		Provider: ftlv1.SecretProvider_SECRET_INLINE.Enum(),
-		Ref:      configRef,
-		Value:    buffer,
-	}))
-	assert.EqualError(t, err, expectedError)
 }
