--- backend/controller/admin/testdata/go/dischema/dischema.go
@@ -1,23 +1,0 @@
-package dischema
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/TBD54566975/ftl/go-runtime/ftl" // Import the FTL SDK.
-)
-
-var defaultUser = ftl.Config[string]("defaultUser")
-
-type EchoRequest struct {
-	Name ftl.Option[string] `json:"name"`
-}
-
-type EchoResponse struct {
-	Message string `json:"message"`
-}
-
-//ftl:verb
-func Echo(ctx context.Context, req EchoRequest) (EchoResponse, error) {
-	return EchoResponse{Message: fmt.Sprintf("Hello, %s!", req.Name.Default(defaultUser.Get(ctx)))}, nil
-}
