--- backend/controller/admin/testdata/go/dischema/ftl.toml
@@ -1,2 +1,0 @@
-module = "dischema"
-language = "go"
