--- backend/controller/admin/testdata/go/dischema/go.mod
@@ -1,43 +1,0 @@
-module ftl/dischema
-
-go 1.22.2
-
-toolchain go1.22.4
-
-require github.com/TBD54566975/ftl v1.1.5
-
-require (
-	connectrpc.com/connect v1.16.1 // indirect
-	connectrpc.com/grpcreflect v1.2.0 // indirect
-	connectrpc.com/otelconnect v0.7.0 // indirect
-	github.com/alecthomas/concurrency v0.0.2 // indirect
-	github.com/alecthomas/participle/v2 v2.1.1 // indirect
-	github.com/alecthomas/types v0.16.0 // indirect
-	github.com/alessio/shellescape v1.4.2 // indirect
-	github.com/danieljoos/wincred v1.2.0 // indirect
-	github.com/go-logr/logr v1.4.2 // indirect
-	github.com/go-logr/stdr v1.2.2 // indirect
-	github.com/godbus/dbus/v5 v5.1.0 // indirect
-	github.com/jackc/pgpassfile v1.0.0 // indirect
-	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
-	github.com/jackc/pgx/v5 v5.6.0 // indirect
-	github.com/jackc/puddle/v2 v2.2.1 // indirect
-	github.com/jpillora/backoff v1.0.0 // indirect
-	github.com/kballard/go-shellquote v0.0.0-20180428030007-95032a82bc51 // indirect
-	github.com/mattn/go-isatty v0.0.20 // indirect
-	github.com/multiformats/go-base36 v0.2.0 // indirect
-	github.com/swaggest/jsonschema-go v0.3.70 // indirect
-	github.com/swaggest/refl v1.3.0 // indirect
-	github.com/zalando/go-keyring v0.2.4 // indirect
-	go.opentelemetry.io/otel v1.27.0 // indirect
-	go.opentelemetry.io/otel/metric v1.27.0 // indirect
-	go.opentelemetry.io/otel/trace v1.27.0 // indirect
-	golang.org/x/crypto v0.23.0 // indirect
-	golang.org/x/exp v0.0.0-20240506185415-9bf2ced13842 // indirect
-	golang.org/x/mod v0.17.0 // indirect
-	golang.org/x/net v0.25.0 // indirect
-	golang.org/x/sync v0.7.0 // indirect
-	golang.org/x/sys v0.20.0 // indirect
-	golang.org/x/text v0.15.0 // indirect
-	google.golang.org/protobuf v1.34.1 // indirect
-)
