--- backend/controller/admin/testdata/go/ftl-project-dr.toml
@@ -1,9 +1,0 @@
-name = "dr"
-ftl-min-version = "dev"
-hermit = false
-no-git = false
-
-[global]
-
-[commands]
-  startup = ["echo disk retreiver"]
