+++ backend/controller/dal/dal.go
@@ -841,23 +841,6 @@ func (d *DAL) GetActiveDeployments(ctx context.Context) ([]Deployment, error) {
 	})
 }
 
-// GetActiveSchema returns the schema for all active deployments.
-func (d *DAL) GetActiveSchema(ctx context.Context) (*schema.Schema, error) {
-	deployments, err := d.GetActiveDeployments(ctx)
-	if err != nil {
-		return nil, err
-	}
-	sch, err := schema.ValidateSchema(&schema.Schema{
-		Modules: slices.Map(deployments, func(d Deployment) *schema.Module {
-			return d.Schema
-		}),
-	})
-	if err != nil {
-		return nil, fmt.Errorf("could not validate schema: %w", err)
-	}
-	return sch, nil
-}
-
 func (d *DAL) GetDeploymentsWithMinReplicas(ctx context.Context) ([]Deployment, error) {
 	rows, err := d.db.GetDeploymentsWithMinReplicas(ctx)
 	if err != nil {
