+++ common/projectconfig/integration_test.go
@@ -3,6 +3,8 @@
 package projectconfig
 
 import (
+	"fmt"
+	"os"
 	"path/filepath"
 	"testing"
 
@@ -12,6 +14,27 @@ import (
 	"github.com/TBD54566975/ftl/internal/exec"
 	"github.com/TBD54566975/ftl/internal/log"
 )
+
+func TestCmdsCreateProjectTomlFilesIfNonexistent(t *testing.T) {
+	cwd, err := os.Getwd()
+	assert.NoError(t, err)
+
+	fileName := "ftl-project-nonexistent.toml"
+	configPath := filepath.Join(cwd, "testdata", "go", fileName)
+
+	in.Run(t, fileName,
+		in.CopyModule("echo"),
+		in.Exec("ftl", "config", "set", "key", "--inline", "value"),
+		in.FileContains(configPath, "key"),
+		in.FileContains(configPath, "InZhbHVlIg"),
+	)
+
+	// The FTL config path is special-cased to use the testdata directory
+	// instead of tmpDir, so we need to clean it up manually.
+	fmt.Printf("Removing config file %s\n", configPath)
+	err = os.Remove(configPath)
+	assert.NoError(t, err)
+}
 
 func TestDefaultToRootWhenModuleDirsMissing(t *testing.T) {
 	in.Run(t, "no-module-dirs-ftl-project.toml",
